Tileset testing

+++ Assets/Scripts/Enums/FaceStateEnum.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-public enum FaceStateEnum : byte
+public enum FaceStateEnum : int
 {
     Empty,
     Filled,
